+++ changelog.md
@@ -62,6 +62,8 @@ Proof-of-concept version.
 ## 1.1.6 (open)
 
 * Dependencies are listed in naft generator
+* Ingredients created in a recipe always have that recipe has owner
+* Chai functionality for add_file and add_key_value
 
 ## Next